feat: add initial project setup with architecture and configuration
A new project has been created with a partially configured architecture.

From the test task, the MVVM architecture was selected, which is implemented here as a modular Clean Architecture.

The architecture already includes a ModuleBuilder that loads the Movies module with its coordinator and factory. View controllers and view models have also been prepared to start developing the functionality of the module.

The Data and Domain layers have been implemented, although they will likely be updated during development.

In the application's Core layer, the following have been implemented:

- A DependencyContainer,

- Constants,

- The main network service of the application,

- A Storage layer with an initial implementation of Core Data, prepared for caching data for the MoviesScene list.

Additionally, some localization has been added, which may be needed during development but could be revised in the final version of the Movies module.

The project architecture is now fully prepared for the implementation of the Movies module.